Refactored structure with inputs/outputs.
