+++ src/qe_tools/extractors/__init__.py
@@ -2,7 +2,7 @@
 from pathlib import Path
 
 from qe_tools.exceptions import PathIsNotAFile
-from qe_tools.parsers.inputs import CpInputFile, PwInputFile
+from qe_tools.inputs import CpInputFile, PwInputFile
 
 SUPPORTED_PARSERS = [
     'PW',
