+++ src/qe_tools/inputs/__init__.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+
+from qe_tools.inputs.cp import CpInputFile
+from qe_tools.inputs.pw import PwInputFile
+
+__all__ = ('PwInputFile', 'CpInputFile')
