+++ src/qe_tools/inputs/utils.py
