--- src/qe_tools/parsers/inputs/__init__.py
@@ -1,6 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from qe_tools.parsers.inputs.cp import CpInputFile
-from qe_tools.parsers.inputs.pw import PwInputFile
-
-__all__ = ('PwInputFile', 'CpInputFile')
